Fix SwiftLint violations and reduce DayDetailView file length
- Move @Environment attributes to their own lines per SwiftLint rules
- Fix @ViewBuilder placement on computed properties (same line as var)
- Extract weather card to DayWeatherGroupBox.swift to reduce file length
- DayDetailView now 468 lines (was 545, limit 500)
